# product-service/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
    "/products": {
      "get": {
        "operationId": "ProductController_getProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Product"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Product"
        ]
      }
    },
    "/products/{id}": {
      "get": {
        "operationId": "ProductController_getProductsById",
        "summary": "Get single Product by id",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Product"
                }
              }
            }
          },
          "404": {
            "description": "Product was not found"
          }
        },
        "tags": [
          "Product"
        ]
      }
    }
  },
  "info": {
    "title": "Shop",
    "description": "Backend shop",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "Product": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "image": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "price",
          "image"
        ]
      }
    }
  }
}